Conflicts:
	build/scripts_for_patch/post_install.php
	sugarcrm/ModuleInstall/PackageManager/tpls/PackageManagerScripts.tpl
	sugarcrm/include/DetailView/DetailView.tpl
	sugarcrm/include/EditView/EditView.tpl
	sugarcrm/include/Expressions/Expression/Parser/Parser.php
	sugarcrm/include/MVC/View/SugarView.php
	sugarcrm/include/MVC/View/tpls/Importvcard.tpl
	sugarcrm/include/MVC/View/views/view.spot.php
	sugarcrm/include/MySugar/MySugar.php
	sugarcrm/include/MySugar/tpls/MySugar.tpl
	sugarcrm/include/SugarCharts/SugarChart.php
	sugarcrm/include/SugarCharts/SugarFlash/SugarFlash.php
	sugarcrm/include/SugarCharts/SugarFlash/SugarFlashReports.php
	sugarcrm/include/SugarObjects/LanguageManager.php
	sugarcrm/include/SugarObjects/templates/file/metadata/quickcreatedefs.php
	sugarcrm/include/SugarTheme/SugarTheme.php
	sugarcrm/include/dir_inc.php
	sugarcrm/include/generic/SugarWidgets/SugarWidgetSubPanelTopScheduleMeetingButton.php
	sugarcrm/include/javascript/sugar_3.js
	sugarcrm/include/templates/TemplateDragDropChooser.php
	sugarcrm/include/utils/file_utils.php
	sugarcrm/install/install_utils.php
	sugarcrm/jssource/JSGroupings.php
	sugarcrm/modules/Administration/DiagnosticDownload.php
	sugarcrm/modules/Administration/RebuildJSLang.php
	sugarcrm/modules/Administration/templates/ConfigureTabs.tpl
	sugarcrm/modules/Administration/templates/GlobalSearchSettings.tpl
	sugarcrm/modules/Administration/templates/ShortcutBar.tpl
	sugarcrm/modules/Administration/templates/enableWirelessModules.tpl
	sugarcrm/modules/Administration/templates/themeSettings.tpl
	sugarcrm/modules/Configurator/Configurator.php
	sugarcrm/modules/Configurator/UploadFileCheck.php
	sugarcrm/modules/DCEReports/controller.php
	sugarcrm/modules/DynamicFields/templates/Fields/TemplateRange.php
	sugarcrm/modules/Emails/Compose.php
	sugarcrm/modules/Home/UnifiedSearchAdvanced.php
	sugarcrm/modules/Home/UnifiedSearchAdvancedForm.tpl
	sugarcrm/modules/Import/views/view.last.php
	sugarcrm/modules/KBDocuments/SearchHome.html
	sugarcrm/modules/KBDocuments/tpls/advancedTab.tpl
	sugarcrm/modules/ModuleBuilder/controller.php
	sugarcrm/modules/Reports/Report.php
	sugarcrm/modules/Teams/TeamSet.php
	sugarcrm/modules/UpgradeWizard/commit.php
	sugarcrm/modules/UpgradeWizard/silentUpgrade_step2.php
	sugarcrm/modules/UpgradeWizard/uw_utils.php
	sugarcrm/modules/WorkFlowActionShells/CreateStep1.php
	sugarcrm/modules/WorkFlowActionShells/CreateStep2.php
	sugarcrm/service/v3/SugarWebServiceImplv3.php
	sugarcrm/service/v4/SugarWebServiceImplv4.php
	sugarcrm/tests/SugarTestImportUtilities.php
	sugarcrm/tests/modules/Home/UnifiedSearchAdvancedTest.php
	sugarcrm/tests/modules/Import/ImportFileTest.php
	sugarcrm/tests/modules/UpgradeWizard/Bug36845Test.php

+++ sugarcrm/modules/UpgradeWizard/uw_utils.php
@@ -593,12 +593,17 @@ function deleteCache(){
 	if(is_dir($cachedir)){
 		$allModFiles = array();
 		$allModFiles = findAllFiles($cachedir,$allModFiles);
-	   foreach($allModFiles as $file){
-	       	if(file_exists($file)){
-				unlink($file);
+		foreach($allModFiles as $file) {
+	       	if(file_exists($file)) {
+	       		if(is_dir($file)) {
+				  rmdir_recursive($file);
+	       		} else {
+	       		  unlink($file);
+	       		}
 	       	}
-	   }
+		}
 	}
+	
 	//Clean jsLanguage from cache
 	$cachedir = sugar_cached('jsLanguage');
 	if(is_dir($cachedir)){
@@ -4265,10 +4270,10 @@ function upgradeDashletsForSalesAndMarketing() {
         'mypbss_date_end' => 'MyPipelineBySalesStageDashlet',
         'mypbss_sales_stages' => 'MyPipelineBySalesStageDashlet',
         'mypbss_chart_type' => 'MyPipelineBySalesStageDashlet',
-        'lsbo_lead_sources' => 'OppByLeadOutcomeDashlet',
-        'lsbo_ids' => 'OppByLeadOutcomeDashlet',
-        'pbls_lead_sources' => 'OppByLeadSourceDashlet',
-        'pbls_ids' => 'OppByLeadSourceDashlet',
+        'lsbo_lead_sources' => 'OpportunitiesByLeadSourceByOutcomeDashlet',
+        'lsbo_ids' => 'OpportunitiesByLeadSourceByOutcomeDashlet',
+        'pbls_lead_sources' => 'OpportunitiesByLeadSourceDashlet',
+        'pbls_ids' => 'OpportunitiesByLeadSourceDashlet',
         'pbss_date_start' => 'PipelineBySalesStageDashlet',
         'pbss_date_end' => 'PipelineBySalesStageDashlet',
         'pbss_sales_stages' => 'PipelineBySalesStageDashlet',
@@ -4836,14 +4841,16 @@ function upgradeTeamColumn($bean, $column_name) {
 			$GLOBALS['db']->addColumn($bean->table_name, $bean->field_defs['team_set_id']);
 		}
 		$indexArray =  $GLOBALS['db']->helper->get_indices($bean->table_name);
-		$indexDef = array(
+		
+        $indexName = getValidDBName('idx_'.strtolower($bean->table_name).'_tmst_id', true, 34);
+        $indexDef = array(
 					 array(
-						'name' => 'idx_'.strtolower($bean->table_name).'_tmst_id',
+						'name' => $indexName,
 						'type' => 'index',
 						'fields' => array('team_set_id')
 					 )
 				   );
-		if(!isset($indexArray['idx_'.strtolower($bean->table_name).'_tmst_id'])) {
+		if(!isset($indexArray[$indexName])) {
 			$GLOBALS['db']->addIndexes($bean->table_name, $indexDef);
 		}
 
@@ -4949,36 +4956,19 @@ function upgradeModulesForTeam() {
             //new modules list now has left over modules which are new to this install, so lets add them to the system tabs
             logThis('new modules to add are '.var_export($newModuleList,true),$path);
 
-            //grab the existing system tabs
-            $tabs = $newTB->get_system_tabs();
-
-            //add the new tabs to the array
-            foreach($newModuleList as $nm ){
-              $tabs[$nm] = $nm;
+            if(!empty($newModuleList))
+            {
+	            //grab the existing system tabs
+	            $tabs = $newTB->get_system_tabs();
+	
+	            //add the new tabs to the array
+	            foreach($newModuleList as $nm ){
+	              $tabs[$nm] = $nm;
+	            }
+	
+	            $newTB->set_system_tabs($tabs);
             }
-
-            if(!file_exists('modules/iFrames/iFrame.php') && isset($tabs['iFrames'])){
-                unset($tabs['iFrames']);
-            }
-
-	        //Set the default order
-	        $default_order = array(
-	        	'Home'=>'Home',
-	        	'Accounts'=>'Accounts',
-	        	'Contacts'=>'Contacts',
-	        	'Opportunities'=>'Opportunities',
-	        	'Activities'=>'Activities',
-	            //BEGIN SUGARCRM flav=pro || flav=sales ONLY
-	        	'Reports'=>'Reports',
-	            //END SUGARCRM flav=pro || flav=sales ONLY
-	        	'Documents'=>'Documents'
-	        );
-	        $tabs = array_merge($default_order, $tabs);
-
-            //now assign the modules to system tabs
-            $newTB->set_system_tabs($tabs);
             logThis('module tabs updated',$path);
-
         }
     }
 
@@ -5636,6 +5626,33 @@ function upgrade_connectors($path='') {
     logThis('End upgrade_connectors', $path);
 }
 
+function repair_long_relationship_names($path='')
+{
+    logThis("Begin repair_long_relationship_names", $path);
+    require_once 'modules/ModuleBuilder/parsers/relationships/DeployedRelationships.php' ;
+    $GLOBALS['mi_remove_tables'] = false;
+    $touched = array();
+    foreach($GLOBALS['moduleList'] as $module)
+    {
+        $relationships = new DeployedRelationships ($module) ;
+        foreach($relationships->getRelationshipList() as $rel_name)
+        {
+            if (strlen($rel_name) > 27 && empty($touched[$rel_name]))
+            {
+                logThis("Rebuilding relationship fields for $rel_name", $path);
+                $touched[$rel_name] = true;
+                $rel_obj = $relationships->get($rel_name);
+                $rel_obj->setReadonly(false);
+                $relationships->delete($rel_name);
+                $relationships->save();
+                $relationships->add($rel_obj);
+                $relationships->save();
+                $relationships->build () ;
+            }
+        }
+    }
+    logThis("End repair_long_relationship_names", $path);
+}
 
 function removeSilentUpgradeVarsCache(){
     global $silent_upgrade_vars_loaded;
@@ -5817,4 +5834,111 @@ function upgradeDisplayedTabsAndSubpanels($version)
 	    $tc->set_system_tabs($tabs[0]);    
 	    logThis('finish upgrading system displayed tabs and subpanels'); 
 	}
+}
+
+
+/**
+ * unlinkUpgradeFiles
+ * This is a helper function to clean up 
+ * 
+ * @param $version String value of current system version (pre upgrade)
+ */
+function unlinkUpgradeFiles($version)
+{
+	if(!isset($version))
+	{
+	   return;
+	}
+	
+	logThis('start unlinking files from previous upgrade');
+	if($version < '620')
+	{
+	   //list of files to remove
+	   $files_to_remove = array('modules/Notifications/metadata/studio.php', 'modules/Help/Forms.php','themes/Sugar5/images/sugarColors.xml');
+	   
+	   foreach($files_to_remove as $f)
+	   {
+		   if(file_exists($f))
+		   {
+		   	  logThis('removing file: ' . $f);
+		   	  unlink($f);
+		   }  
+	   }
+	}
+	logThis('end unlinking files from previous upgrade');
+	
+	if($version < '620')
+	{
+		logThis('start upgrade for DocumentRevisions classic files (EditView.html, EditView.php, DetailView.html, DetailView.php)');
+
+		//Use a md5 comparison check to see if we can just remove the file where an exact match is found
+		if($version < '610')
+		{
+			$dr_files = array(
+	         'modules/DocumentRevisions/DetailView.html' => '17ad4d308ce66643fdeb6fdb3b0172d3',
+			 'modules/DocumentRevisions/DetailView.php' => 'd8606cdcd0281ae9443b2580a43eb5b3',
+	         'modules/DocumentRevisions/EditView.php' => 'c7a1c3ef2bb30e3f5a11d122b3c55ff1',
+	         'modules/DocumentRevisions/EditView.html' => '7d360ca703863c957f40b3719babe8c8',
+	        );		
+		} else {
+			$dr_files = array(
+	         'modules/DocumentRevisions/DetailView.html' => 'a8356ff20cd995daffe6cb7f7b8b2340',
+			 'modules/DocumentRevisions/DetailView.php' => '20edf45dd785469c484fbddff1a3f8f2',
+	         'modules/DocumentRevisions/EditView.php' => 'fb31958496f04031b2851dcb4ce87d50',
+	         'modules/DocumentRevisions/EditView.html' => 'b8cada4fa6fada2b4e4928226d8b81ee',
+	        );
+		}
+	
+		foreach($dr_files as $rev_file=>$hash)
+		{
+			if(file_exists($rev_file))
+			{
+				//It's a match here so let's just remove the file
+				if (md5(file_get_contents($rev_file)) == $hash) 
+				{
+					logThis('removing file ' . $rev_file);
+					unlink($rev_file);
+				} else {
+					if(!copy($rev_file, $rev_file . '.suback.bak')) 
+					{
+					  logThis('error making backup for file ' . $rev_file);
+					} else {
+					  logThis('copied file ' . $rev_file . ' to ' . $rev_file . '.suback.bak');
+					  unlink($rev_file);
+					}
+				} 
+			}
+		}
+		
+		logThis('end upgrade for DocumentRevisions classic files');
+	}	
+}
+
+if (!function_exists("getValidDBName"))
+{
+    /*
+     * Return a version of $proposed that can be used as a column name in any of our supported databases
+     * Practically this means no longer than 25 characters as the smallest identifier length for our supported DBs is 30 chars for Oracle plus we add on at least four characters in some places (for indicies for example)
+     * @param string $name Proposed name for the column
+     * @param string $ensureUnique
+     * @return string Valid column name trimmed to right length and with invalid characters removed
+     */
+     function getValidDBName ($name, $ensureUnique = false, $maxLen = 30)
+    {
+        // first strip any invalid characters - all but alphanumerics and -
+        $name = preg_replace ( '/[^\w-]+/i', '', $name ) ;
+        $len = strlen ( $name ) ;
+        $result = $name;
+        if ($ensureUnique)
+        {
+            $md5str = md5($name);
+            $tail = substr ( $name, -11) ;
+            $temp = substr($md5str , strlen($md5str)-4 );
+            $result = substr ( $name, 0, 10) . $temp . $tail ;
+        }else if ($len > ($maxLen - 5))
+        {
+            $result = substr ( $name, 0, 11) . substr ( $name, 11 - $maxLen + 5);
+        }
+        return strtolower ( $result ) ;
+    }
 }
